requirements.txt
scrapy
numpy
pandas
beautifulsoup4
ipython
python-dotenv
langchain
langchain_openai
langchain_chroma
openai
streamlit